requirements.txt
torch
numpy
pillow
torchvision
matplotlib
tqdm
tensorboard
pyyaml